Spreadsheet of systems from Wookieepedia added
Note: 3775 planets seems low given that 1800+ systems have been added to the Outer Rim Territories alone.  There are many systems that are described as being inhabited but have no known planets, inhabited or otherwise.  Perhaps this is causing the discrepancy.  Regardless, this is not as helpful of a metric for progress as I had hoped.